+++ frontend/src/components/Toast.tsx
@@ -1,21 +1,21 @@
-import { Toast, ToastContainer } from "react-bootstrap";
-import { useTranslation } from "react-i18next";
-import { Link } from "react-router-dom";
-import "@/styles/products/Toast.css";
+import { Toast, ToastContainer } from "react-bootstrap"
+import { useTranslation } from "react-i18next"
+import { Link } from "react-router-dom"
+import "@/styles/products/Toast.css"
 
 interface ToastNotificationProps {
   toasts: {
-    id: number;
-    message: string;
-    image: string;
-    name: string;
-    type?: "success" | "danger"; // Optional type, defaults to success
-  }[];
-  onClose: (id: number) => void;
+    id: number
+    message: string
+    image: string
+    name: string
+    type?: "success" | "danger" // Optional type, defaults to success
+  }[]
+  onClose: (id: number) => void
 }
 
 const ToastNotification = ({ toasts, onClose }: ToastNotificationProps) => {
-  const { t } = useTranslation();
+  const { t } = useTranslation()
 
   return (
     <ToastContainer className="toast-container">
@@ -31,14 +31,16 @@ const ToastNotification = ({ toasts, onClose }: ToastNotificationProps) => {
         >
           <Toast.Header closeButton className="toast-header">
             <strong className="me-auto">
-              {toast.type === "danger" ? t("cart.remove.title") : t("cart.add.title")}
+              {toast.type === "danger"
+                ? t("cart.remove.title")
+                : t("cart.add.title")}
             </strong>
           </Toast.Header>
           <Toast.Body className="toast-body">
             <img src={toast.image} alt={toast.name} />
             <span>{toast.message}</span>
             {toast.type !== "danger" && (
-              <Link to="/lang/cart" className="btn btn-success mt-2">
+              <Link to="/lang/cart" className="btn btn-success m-2">
                 {t("cart.showCart")}
               </Link>
             )}
@@ -46,7 +48,7 @@ const ToastNotification = ({ toasts, onClose }: ToastNotificationProps) => {
         </Toast>
       ))}
     </ToastContainer>
-  );
-};
+  )
+}
 
-export default ToastNotification;
+export default ToastNotification