+++ frontend/src/pages/About.tsx
@@ -1,8 +1,8 @@
-import { useTranslation } from "react-i18next";
-import { Container } from "react-bootstrap";
+import { useTranslation } from "react-i18next"
+import { Container } from "react-bootstrap"
 
 const About = () => {
-  const { t } = useTranslation();
+  const { t } = useTranslation()
   return (
     <Container fluid className="d-flex flex-column min-vh-100 p-0">
       <h2 className="mb-4">{t("about")}</h2>
@@ -11,7 +11,7 @@ const About = () => {
         <p className="lead">{t("aboutText.second")}</p>
       </div>
     </Container>
-  );
-};
+  )
+}
 
-export default About;
+export default About